Rename component files to use PascalCase.
