+++ app/components/QuoteCard.jsx
