+++ app/components/popup.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import LoginSignup from './login-signup';
-import FinishAccountCreation from './finish-account-creation';
-import CreateBoard from './create-board';
+import Login from './Login';
+import FinishAccountCreation from './FinishAccountCreation';
+import CreateBoard from './CreateBoard';
 
 export default class Popup extends React.Component {
 
@@ -9,7 +9,7 @@ export default class Popup extends React.Component {
     let content = null;
     switch (this.props.type) {
       case 'login':
-        content = <LoginSignup {...this.props}/>;
+        content = <Login {...this.props}/>;
         break;
       case 'finishAccountCreation':
         content = <FinishAccountCreation {...this.props} />;
